chore: move tasks in oscar_to_econvention.py fix: remove f-string SQL request to prevent SQL injection
